requirements.txt
appnope==0.1.0
attrs==17.4.0
backcall==0.1.0
boto3==1.7.2
botocore==1.10.2
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
dj-database-url==0.5.0
Django==2.0.4
django-cors-headers==2.2.0
django-extensions==2.0.6
-e git+https://github.com/fernandoe/django-fe-core.git@release/0.4.0#egg=django_fe_core
-e git+https://github.com/fernandoe/django-fe-jwt.git@release/0.0.1#egg=django_fe_jwt
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
factory-boy==2.10.0
Faker==0.8.12
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
jmespath==0.9.3
more-itertools==4.1.0
mysqlclient==1.3.12
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
PyJWT==1.6.1
pytest==3.5.0
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
